+++ packages/data/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @apex-stack/data
 
+## 0.8.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @apex-stack/core@0.20.0
+
 ## 0.8.0
 
 ### Minor Changes